% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisEdad.R
\name{AnalisisEdad}
\alias{AnalisisEdad}
\title{Distribucion por Edad}
\usage{
AnalisisEdad
}
\description{
Esta funcion permite generar una visualizacion interactia con la distribucion por edad de los atletas olimpicos a lo largo de los años.
}
\details{
Esta funcion permite generar una visualizacion interactiva con los datos historicos de los juegos olimpicos desde la edicion de Atenas 1896.
La visualizacion consiste en la distribucion por Edad de los diferentes competidores, ademas se agregaron una serie de filtros
para facilitar la interaccion con la informacion y ademas permitir parametrizar los datos que se desean representar.
}
