% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisGeneral.R
\name{AnalisisGeneral}
\alias{AnalisisGeneral}
\title{Analisis General}
\usage{
AnalisisGeneral()
}
\description{
Esta funcion permite generar una serie de visualizaciones con los datos de los juegos olimpicos.
}
\details{
Esta funcion permite generar 3 visualizaciones interactivas para representar la evolucion de los juegos olimpicos, considerando la cantidad de Atletas participantes, Naciones y Eventos.
}
