% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisGenero.R
\name{AnalisisGenero}
\alias{AnalisisGenero}
\title{Analisis por Genero}
\usage{
AnalisisGenero()
}
\description{
Esta funcion permite representar la participacion de Hombres y Mujeres en los Juegos Olimpicos a lo largo de los años.
}
\details{
Esta funcion permite representar la participacion de Hombres y Mujeres en los Juegos Olimpicos a lo largo de los años.
}
