% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisMapa.R
\name{AnalisisMapa}
\alias{AnalisisMapa}
\title{Mapa Coropletico}
\usage{
AnalisisMapa()
}
\description{
Esta Funcion permite crear un mapa coropletico con la informacion de las medallas ganadas por cada pais
}
\details{
Esta funcion permite generar una visualizacion interactiva con los datos historicos de los juegos olimpicos desde la edicion de Atenas 1896.
La visualizacion consiste en la representacion grafica de la cantidad de medallas ganadas por cada pais.
Se utiliza un mapa coropletico para representar dicha informacion. La visualizacion permtie recorrer un mapa en 2D con la informacion de cada pais.
Ademas se agregaron una serie de filtros que permiten interactuar con la informacion graficada y parametrizar lo que se desea representar.
}
