% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisPesoAltura.R
\name{AnalisisPesoAltura}
\alias{AnalisisPesoAltura}
\title{Peso VS Altura}
\usage{
AnalisisPesoAltura()
}
\description{
Esta funcion permite generar un diagrama de dispersion con la informacion de Peso y Altura de cada Atleta.
}
\details{
Esta funcion permite generar un diagrama de dispersion con la informacion de Peso y Altura de cada Atleta.
}
