% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisPIB.R
\name{AnalisisPIB}
\alias{AnalisisPIB}
\title{Correlacion entre el Producto Interno Bruto (PIB) y la cantidad de medallas ganadas}
\usage{
AnalisisPIB()
}
\description{
Esta funcion permite representar el PIB de cada pais en relacion a la cantidad de medallas ganadas.
}
\details{
Esta funcion permite generar una visualizacion interactiva con los datos historicos de los juegos olimpicos.
La visualizacion consiste en la representacion de la relacion entre el PIB de cada pais y la cantidad de medallas ganadas.
Para comprobar la fuerza de la relacion entre las dos variable se realiza el calculo de la correlacion.
Ademas se agregaron una serie de filtros que permiten interactuar con la informacion graficada y ademas poder parametrizar lo que se desea representar.
}
