% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisTopPaises.R
\name{AnalisisTopPaises}
\alias{AnalisisTopPaises}
\title{TOP Paises}
\usage{
AnalisisTopPaises()
}
\description{
Esta Funcion permite crear una visualizacion interactiva con la informacion de los paises que mas medallas ganan en cada uno de lso deportes.
}
\details{
Esta funcion permite generar una visualizacion interactiva con los datos historicos de los juegos olimpicos desde la edicion de Atenas 1896.
La visualizacion consiste en la representacion del TOP 15 de paises que mas medallan han ganado en cada uno de los deportes.
Sin embargo tambien permite buscar un pais determinado que no este dentro de este top15 para poder visualizar las medallas que han ganado.
Se agregaron una serie de filtros para facilitar la interaccion con la informacion.
}
