% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardOlimpico.R
\name{DashboardOlimpico}
\alias{DashboardOlimpico}
\title{Dashboard Olimpico}
\usage{
DashboardOlimpico()
}
\description{
Esta funcion permite generar un  Dashboard con una serie de visualizaciones sobre los datos historicos de los juegos olimpicos.
}
\details{
Esta funcion permite generar un dashboard interactivos con una serie de visualizaciones y filtros de lso datos historicos de los juegos olimpicos desde la edicion de Atenas 1896.
}
