% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DatosEdad}
\alias{DatosEdad}
\title{18 Olympics variables for Age}
\format{An object of class \code{data.frame} with 261642 rows and 18 columns.}
\usage{
DatosEdad
}
\description{
Variables para edad de los Juegos Olimpicos desde Atenas 1986
}
\keyword{datasets}
