% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DatosOlimpicos}
\alias{DatosOlimpicos}
\title{#' 18 Olympics variables}
\format{An object of class \code{data.frame} with 271116 rows and 18 columns.}
\usage{
DatosOlimpicos
}
\description{
Variables de los Juegos Olimpicos desde Atenas 1986
}
\keyword{datasets}
