% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DatosResumen}
\alias{DatosResumen}
\title{#' 10 variables for Olympics Games}
\format{An object of class \code{data.frame} with 271116 rows and 10 columns.}
\usage{
DatosResumen
}
\description{
Variables Resumen de los Juegos Olimpicos desde Atenas 1986
}
\keyword{datasets}
