% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DatosUnificados}
\alias{DatosUnificados}
\title{#' 6 variables for Olympics Games}
\format{An object of class \code{data.frame} with 812 rows and 6 columns.}
\usage{
DatosUnificados
}
\description{
Variables Unificadas de los Juegos Olimpicos desde Atenas 1986
}
\keyword{datasets}
