% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraficaTabular.R
\name{GraficaTabular}
\alias{GraficaTabular}
\title{Grafica Tabular.}
\usage{
GraficaTabular()
}
\description{
Esta es una funcion para mostrar todos los datos historicos de los juegos olimpicos.
}
\details{
Esta funcion permite mostrar de forma tabular los datos historicos de los juegos olimpicos desde la edicion de Atenas 1896.
Se agregaron filtros en cada una de las columnas de la visualizacion, para que se pueda filtrar la informacion que se necesite.
Ademas se agregaron secciones para descargar todos estos datos historicos, asi como para  imprimirlos y exportarlos a diferentes formatos.
}
