% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosOlimpicos.R
\name{ObtenerDatosOlimpicos}
\alias{ObtenerDatosOlimpicos}
\title{Obtener Datos Historicos de los Juegos Olimpicos}
\usage{
ObtenerDatosOlimpicos()
}
\description{
Esta es una funcion para obtener los datos historicos de los juegos olimpicos
}
\details{
Esta funcion permite obtener los datos de los juegos Olimpicos desde el año 1896
Este Conjunto de Datos se ha procesado previamente (Limpiado, saneado y transformado)
}
